pyproject.toml
[project]
name = "hltb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "fastapi[standard]>=0.121.3",
  "howlongtobeatpy>=1.0.18",
  "python-dotenv>=1.2.1",
  "uvicorn>=0.38.0",
]

